feat: enable ctty support.
This patch make hustmirror support user input when is running in a pipe.

Hustmirror used to use stdin for user input, but when the main script
is piped, for example, when run `curl URL_STORE_THIS_SCRIPT | sh`,
it not work, and what we can do is only to use `test` to judge stdin
is a tty and print corresponding message.

However, the controlling terminal (ctty) can solve this problem, the
`/dev/tty` file is POSIX ctty, and can read from it to test whether
it is available for user's input.

If is not exsist or cannot read, it means user really no not have a
tty, or we can read from /dev/tty.